fix: rename careerStats.ts to careerStatsSelectors.ts to avoid case-insensitive filename collision with CareerStats.tsx
